src/requirements.txt
flask
flask_cors
flask_caching
mysql.connector
dotenv
PyJWT
werkzeug